Create LocalEnumerationModule.psm1
The LocalEnumerationModule is a PowerShell module designed to facilitate the collection of various system and network information from a local Windows system. It provides a collection of functions that leverage common command-line utilities and Windows Management Instrumentation (WMI) queries to retrieve valuable insights about the system's configuration, users, processes, services, network, and more.